% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{filter_df_by_csv}
\alias{filter_df_by_csv}
\title{Filter a DataFrame Based on Conditions from a CSV File}
\usage{
filter_df_by_csv(df, sep, filter_csv_path)
}
\arguments{
\item{df}{A dataframe to be filtered.}

\item{sep}{A character string specifying the field separator character for the CSV file.}

\item{filter_csv_path}{A string containing the path to the CSV file with the filter conditions.}
}
\value{
The filtered dataframe after all conditions have been applied.
}
\description{
This function filters rows in a dataframe by applying conditions specified in a CSV file.
The CSV file should contain two columns: `column_name` and `condition`. For each row in the CSV,
the function applies the specified condition to the specified column and filters the dataframe accordingly.
}
