% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{gather_group_by_summarise}
\alias{gather_group_by_summarise}
\title{Reshape, Group, and Summarise a DataFrame}
\usage{
gather_group_by_summarise(
  df,
  category,
  value,
  column_span,
  group_by_cols,
  summary_funcs
)
}
\arguments{
\item{df}{A dataframe to be reshaped, grouped, and summarised.}

\item{category}{A string specifying the name of the new column that will hold the reshaped variable names (from wide to long format).}

\item{value}{A string specifying the name of the new column that will hold the reshaped values (from wide to long format).}

\item{column_span}{A character vector specifying the columns to reshape from wide to long format.}

\item{group_by_cols}{A character vector specifying the columns to group by after reshaping.}

\item{summary_funcs}{A named list of functions to be applied as summaries to the `value` column after grouping.}
}
\value{
A summarised dataframe with the results of the applied summary functions.
}
\description{
This function reshapes a dataframe from wide to long format, groups by specified columns, and applies multiple summary functions to the values in the reshaped dataframe.
}
