% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{load_and_clean_csv}
\alias{load_and_clean_csv}
\title{Load and Clean CSV Data}
\usage{
load_and_clean_csv(data_path, var_path, sep, header, encode)
}
\arguments{
\item{data_path}{A character string specifying the path to the CSV data file.}

\item{var_path}{A character string specifying the path to the CSV file containing the variable mapping.
This file should have two columns: the first for the new variable names and the second for the old variable names.}

\item{sep}{A character string specifying the field separator character for both CSV files (e.g., `","`, `";"`).}

\item{header}{A logical value indicating whether the CSV files contain a header row.}

\item{encode}{A character string specifying the encoding of the CSV files (e.g., `"UTF-8"`, `"ISO-8859-1"`).}
}
\value{
A cleaned data frame where:
\describe{
  \item{Duplicates}{are removed based on the `Index` column.}
  \item{Selected columns}{are based on the old variable names provided in the variable mapping file.}
  \item{Columns}{are renamed based on the corresponding new variable names.}
  \item{Rows}{containing only missing values are removed.}
}
}
\description{
This function loads a CSV file, selects and renames columns based on a provided variable mapping,
removes duplicate rows based on an index column, and filters out rows with only missing values.
}
