% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_clean_data.R
\name{remove_na_by_columns}
\alias{remove_na_by_columns}
\title{Remove Rows with NA Values from Specified Columns}
\usage{
remove_na_by_columns(df, column_names)
}
\arguments{
\item{df}{A dataframe from which rows with `NA` values will be removed.}

\item{column_names}{A character vector containing the names of the columns to check for `NA` values.}
}
\value{
A dataframe with rows containing `NA` values removed from the specified columns.
}
\description{
This function removes rows from a dataframe that contain `NA` (missing values) in any of the specified columns.
The function iterates through the list of column names and removes rows that have missing values in each specified column.
}
